Parameterize the platform on Unix script to allow creating an iOS app

+++ mobile/app/createapp.sh
@@ -1,4 +1,10 @@
 #!/bin/sh
+PLATFORM=$1
+if [ -x ${PLATFORM} ];
+then
+	echo PLATFORM not defined
+	exit 1
+fi
 rm -rf wallet
 cordova create wallet org.nxt.mobile.wallet "NXT Mobile Wallet" --template ../../html
 cd wallet
@@ -6,7 +12,7 @@ rm -rf icons
 rm -rf plugins
 cp -a ../../icons icons
 cp -a ../../plugins plugins
-cordova platform add android
+cordova platform add ${PLATFORM}
 rm -rf platforms
 cp -a ../../platforms platforms
 cd ..
